Extract CardCoordinates into card_coordinates.dart
