+++ lib/src/sdk/cards/card_models.dart
@@ -2,17 +2,6 @@ import 'package:equatable/equatable.dart' show EquatableMixin;
 
 import '../trello_object.dart' show TrelloObject;
 
-// Field vale may come from Trello as 'latitude,longitude'
-class CardCoordinates with EquatableMixin {
-  late double latitude;
-  late double longitude;
-
-  CardCoordinates({required this.latitude, required this.longitude});
-
-  @override
-  List<Object?> get props => [latitude, longitude];
-}
-
 class CardLabel with EquatableMixin {
   late String id;
   late String idBoard;
