+++ lib/src/sdk/cards/cards.dart
@@ -2,6 +2,7 @@ export 'attachment_client.dart';
 export 'attachment_id.dart';
 export 'card.dart';
 export 'card_client.dart';
+export 'card_coordinates.dart';
 export 'card_id.dart';
 export 'card_models.dart';
 export 'trello_card.dart';
